Update BloodMage.cs
Tried essentially copying AscendedMages way of doing this. It only calls this once now, which is awesome. However, it doesn't seem to fully work, the log values are right when they output, but the health/stamina isnt applied and the well of mana stack isn't removed, but the mana points are. It's all wack.